+++ src/scip/cons_disjunction.c
@@ -666,7 +666,7 @@ SCIP_DECL_CONSPARSE(consParseDisjunction)
    nconss = 0;
    sconss = 10;
    SCIP_CALL( SCIPallocBufferArray(scip, &conss, sconss) );
-   SCIP_CALL( SCIPduplicateBufferArray(scip, &copystr, str, strlen(str)+1) );
+   SCIP_CALL( SCIPduplicateBufferArray(scip, &copystr, str, (int)strlen(str)+1) );
 
    /* find '(' at the beginning, string should start with 'disjunction(' */
    saveptr = strpbrk(copystr, "("); /*lint !e158*/
